requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
auto-py-to-exe==2.43.3
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.2
capstone==5.0.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored-traceback==0.3.0
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
deeppavlov==1.6.0
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.89.1
fastjsonschema==2.19.1
filelock==3.9.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
fqdn==1.5.1
frida==16.2.1
frida-tools==12.3.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keyboard==0.13.5
konlpy==0.6.0
lxml==5.2.1
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
matplotlib-inline==0.1.6
mistune==3.0.2
MouseInfo==0.1.3
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.23.5
openai==1.33.0
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
plumbum==1.8.3
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
pwntools==4.12.0
PyAutoGUI==0.9.54
pybind11==2.10.3
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.18.1
pyelftools==0.31
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ROPGadget==7.4
rpds-py==0.18.0
rpyc==6.0.0
scikit-learn==1.4.0
scipy==1.10.0
selenium==4.21.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.22.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.2
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
unicorn==2.0.1.post1
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.18.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.4
whichcraft==0.6.1
widgetsnbextension==4.0.10
wsproto==1.2.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
zstandard==0.22.0
